Separating the file "file.ts" from the Facebook and WhatsApp channels. Adding the content types and sizes file supported on the WhatsApp channel. Removing the "fileCaption" field of type file from the Facebook channel.
