+++ spec/tags/whatsapp.md
@@ -4,11 +4,22 @@ To activate WhatsApp you a need a registered number on WhatsApp Business API and
 
 **Get in touch with Zenvia consultants to start your account creation.**
 
+
 ## Limitations
 
 The WhatsApp API has some limitations:
 
 * To start a conversation with someone you need a special type of message (a template message) and an approval before being used.
+
+Supported content types and sizes:
+
+| Media | Content Type | Size |
+|---|---|---|
+| document | Any valid MIME type. | 100 MB |
+| image | image/jpeg<br>image/png | 5 MB |
+| audio | audio/acc<br>audio/mp4<br>audio/amr<br>audio/mpeg<br>audio/ogg; codecs=opus | 16 MB |
+| video | video/mp4<br>video/3gpp<br>**Note:** Only H.264 video codec and AAC audio codec is supported. | 16 MB |
+
 
 ## Message sender and recipient
